+++ provers/sp1/driver/src/distributed/mod.rs
@@ -1,0 +1,4 @@
+mod prover;
+mod worker;
+
+pub use prover::Sp1DistributedProver;
